[fli-iam#248] Manage size property to dataset expression at import level
