+++ shanoir-ng-datasets/src/main/java/org/shanoir/ng/importer/service/ImporterService.java
@@ -14,6 +14,35 @@
 
 package org.shanoir.ng.importer.service;
 
+import org.shanoir.ng.dataset.modality.*;
+import org.shanoir.ng.dataset.model.Dataset;
+import org.shanoir.ng.dataset.model.*;
+import org.shanoir.ng.dataset.service.DatasetService;
+import org.shanoir.ng.datasetacquisition.model.DatasetAcquisition;
+import org.shanoir.ng.datasetacquisition.service.DatasetAcquisitionService;
+import org.shanoir.ng.datasetfile.DatasetFile;
+import org.shanoir.ng.examination.model.Examination;
+import org.shanoir.ng.examination.repository.ExaminationRepository;
+import org.shanoir.ng.examination.service.ExaminationService;
+import org.shanoir.ng.importer.dto.*;
+import org.shanoir.ng.processing.model.DatasetProcessing;
+import org.shanoir.ng.shared.event.ShanoirEvent;
+import org.shanoir.ng.shared.event.ShanoirEventService;
+import org.shanoir.ng.shared.event.ShanoirEventType;
+import org.shanoir.ng.shared.exception.ShanoirException;
+import org.shanoir.ng.solr.service.SolrService;
+import org.shanoir.ng.utils.KeycloakUtil;
+import org.shanoir.ng.utils.SecurityContextUtil;
+import org.shanoir.ng.utils.Utils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Scope;
+import org.springframework.mock.web.MockMultipartFile;
+import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
+
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
@@ -21,82 +50,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-
-import org.shanoir.ng.dataset.modality.CalibrationDataset;
-import org.shanoir.ng.dataset.modality.CtDataset;
-import org.shanoir.ng.dataset.modality.EegDataset;
-import org.shanoir.ng.dataset.modality.EegDatasetDTO;
-import org.shanoir.ng.dataset.modality.MegDataset;
-import org.shanoir.ng.dataset.modality.MeshDataset;
-import org.shanoir.ng.dataset.modality.MrDataset;
-import org.shanoir.ng.dataset.modality.ParameterQuantificationDataset;
-import org.shanoir.ng.dataset.modality.PetDataset;
-import org.shanoir.ng.dataset.modality.ProcessedDatasetType;
-import org.shanoir.ng.dataset.modality.RegistrationDataset;
-import org.shanoir.ng.dataset.modality.SegmentationDataset;
-import org.shanoir.ng.dataset.modality.SpectDataset;
-import org.shanoir.ng.dataset.modality.StatisticalDataset;
-import org.shanoir.ng.dataset.modality.TemplateDataset;
-import org.shanoir.ng.dataset.model.CardinalityOfRelatedSubjects;
-import org.shanoir.ng.dataset.model.Dataset;
-import org.shanoir.ng.dataset.model.DatasetExpression;
-import org.shanoir.ng.dataset.model.DatasetExpressionFormat;
-import org.shanoir.ng.dataset.model.DatasetMetadata;
-import org.shanoir.ng.dataset.model.DatasetModalityType;
-import org.shanoir.ng.dataset.service.DatasetService;
-import org.shanoir.ng.datasetacquisition.model.DatasetAcquisition;
-import org.shanoir.ng.datasetacquisition.model.eeg.EegDatasetAcquisition;
-import org.shanoir.ng.datasetacquisition.service.DatasetAcquisitionService;
-import org.shanoir.ng.datasetfile.DatasetFile;
-import org.shanoir.ng.eeg.model.Channel;
-import org.shanoir.ng.eeg.model.Channel.ChannelType;
-import org.shanoir.ng.eeg.model.Event;
-import org.shanoir.ng.examination.model.Examination;
-import org.shanoir.ng.examination.repository.ExaminationRepository;
-import org.shanoir.ng.examination.service.ExaminationService;
-import org.shanoir.ng.importer.dto.EegImportJob;
-import org.shanoir.ng.importer.dto.ImportJob;
-import org.shanoir.ng.importer.dto.Patient;
-import org.shanoir.ng.importer.dto.ProcessedDatasetImportJob;
-import org.shanoir.ng.importer.dto.Serie;
-import org.shanoir.ng.importer.dto.Study;
-import org.shanoir.ng.processing.model.DatasetProcessing;
-import org.shanoir.ng.shared.configuration.RabbitMQConfiguration;
-import org.shanoir.ng.shared.email.EmailBase;
-import org.shanoir.ng.shared.email.EmailDatasetImportFailed;
-import org.shanoir.ng.shared.email.EmailDatasetsImported;
-import org.shanoir.ng.shared.event.ShanoirEvent;
-import org.shanoir.ng.shared.event.ShanoirEventService;
-import org.shanoir.ng.shared.event.ShanoirEventType;
-import org.shanoir.ng.shared.exception.ShanoirException;
-import org.shanoir.ng.solr.service.SolrService;
-import org.shanoir.ng.study.rights.StudyUser;
-import org.shanoir.ng.study.rights.StudyUserRightsRepository;
-import org.shanoir.ng.utils.KeycloakUtil;
-import org.shanoir.ng.utils.SecurityContextUtil;
-import org.shanoir.ng.utils.Utils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.amqp.AmqpException;
-import org.springframework.amqp.rabbit.core.RabbitTemplate;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Scope;
-import org.springframework.mock.web.MockMultipartFile;
-import org.springframework.stereotype.Service;
-import org.springframework.util.CollectionUtils;
-import org.springframework.web.multipart.MultipartFile;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import java.util.*;
 
 @Service
 @Scope("prototype")
@@ -131,22 +85,12 @@ public class ImporterService {
 	private ShanoirEventService eventService;
 
 	@Autowired
-	StudyUserRightsRepository studyUserRightRepo;
-
+	private SolrService solrService;
 	@Autowired
-	RabbitTemplate rabbitTemplate;
-
-	@Autowired
-	SolrService solrService;
-	
-	@Autowired
-	private ObjectMapper objectMapper;
-
-	private static final String SESSION_PREFIX = "ses-";
+	private ImporterMailService mailService;
 
 	private static final String SUBJECT_PREFIX = "sub-";
 
-	private static final String EEG_PREFIX = "eeg";
 	
 	private static final String PROCESSED_DATASET_PREFIX = "processed-dataset";
 
@@ -162,7 +106,7 @@ public class ImporterService {
     }
 
 	public void createAllDatasetAcquisition(ImportJob importJob, Long userId) throws ShanoirException {
-		LOG.info("createAllDatasetAcquisition: " + this.toString() + " instances: " + getInstancesCreated());
+		LOG.info("createAllDatasetAcquisition: " + this + " instances: " + getInstancesCreated());
 		ShanoirEvent event = importJob.getShanoirEvent();
 		event.setMessage("Creating datasets...");
 		eventService.publishEvent(event);
@@ -226,7 +170,7 @@ public class ImporterService {
 			}
 
 			// Send success mail
-			sendImportEmail(importJob, userId, examination, generatedAcquisitions);
+			mailService.sendImportEmail(importJob, userId, examination, generatedAcquisitions);
 
 		} catch (Exception e) {
 			event.setStatus(ShanoirEvent.ERROR);
@@ -236,86 +180,9 @@ public class ImporterService {
 			LOG.error("Error during import for exam: {} : {}", importJob.getExaminationId(), e);
 			
 			// Send mail
-			sendFailureMail(importJob, userId, e.getMessage());
+			mailService.sendFailureMail(importJob, userId, e.getMessage());
 			
 			throw new ShanoirException(event.getMessage(), e);
-		}
-	}
-
-	/**
-	 * Sens the import email through rabbitMQ to user MS
-	 * @param importJob the import job
-	 * @param userId the userID
-	 * @param examination the exam ID
-	 * @param generatedAcquisitions
-	 */
-	private void sendImportEmail(ImportJob importJob, Long userId, Examination examination, Set<DatasetAcquisition> generatedAcquisitions) {
-		EmailDatasetsImported generatedMail = new EmailDatasetsImported();
-
-		Map<Long, String> datasets = new HashMap<>();
-		if (CollectionUtils.isEmpty(generatedAcquisitions)) {
-			return;
-		}
-		generatedMail.setExamDate(examination.getExaminationDate().toString());
-		generatedMail.setExaminationId(examination.getId().toString());
-		generatedMail.setStudyId(importJob.getStudyId().toString());
-		generatedMail.setSubjectName(importJob.getSubjectName());
-		generatedMail.setStudyName(importJob.getStudyName());
-		generatedMail.setUserId(userId);
-		generatedMail.setStudyCard(importJob.getStudyCardName());
-
-		for (DatasetAcquisition acq : generatedAcquisitions) {
-			if (!CollectionUtils.isEmpty(acq.getDatasets())) {
-				for (Dataset dataset : acq.getDatasets()) {
-					datasets.put(dataset.getId(), dataset.getName());
-				}
-			}
-		}
-
-		generatedMail.setDatasets(datasets);
-		sendMail(importJob, generatedMail, RabbitMQConfiguration.IMPORT_DATASET_MAIL_QUEUE);
-	}
-
-	private void sendFailureMail(ImportJob importJob, Long userId, String errorMessage) {
-		EmailDatasetImportFailed generatedMail = new EmailDatasetImportFailed();
-		generatedMail.setExaminationId(importJob.getExaminationId().toString());
-		generatedMail.setStudyId(importJob.getStudyId().toString());
-		generatedMail.setSubjectName(importJob.getSubjectName());
-		generatedMail.setStudyName(importJob.getStudyName());
-		generatedMail.setUserId(userId);
-		
-		generatedMail.setErrorMessage(errorMessage != null ? errorMessage : "An unexpected error occured, please contact Shanoir support.");
-
-		sendMail(importJob, generatedMail, RabbitMQConfiguration.IMPORT_DATASET_FAILED_MAIL_QUEUE);
-	}
-
-	/**
-	 * Sends the given mail in entry to all recipients in a given study
-	 * @param job the imprt job
-	 * @param email the recipients
-	 * @param queue 
-	 */
-	private void sendMail(ImportJob job, EmailBase email, String queue) {
-		List<Long> recipients = new ArrayList<>();
-
-		// Get all recpients
-		List<StudyUser> users = (List<StudyUser>) studyUserRightRepo.findByStudyId(job.getStudyId());
-
-		for (StudyUser user : users) {
-			if (user.isReceiveNewImportReport()) {
-				recipients.add(user.getUserId());
-			}
-		}
-		if (recipients.isEmpty()) {
-			// Do not send any mail if no recipients
-			return;
-		}
-		email.setRecipients(recipients);
-
-		try {
-			rabbitTemplate.convertAndSend(queue, objectMapper.writeValueAsString(email));
-		} catch (AmqpException | JsonProcessingException e) {
-			LOG.error("Could not send email for this import. ", e);
 		}
 	}
 	
@@ -378,162 +245,14 @@ public class ImporterService {
 			LOG.error("cleanTempFiles: workFolder is null");
 		}
 	}
-	
-	/**
-	 * Create a dataset acquisition, and associated dataset.
-	 * @param importJob the import job from importer MS.
-	 */
-	public void createEegDataset(final EegImportJob importJob) {
-
-		Long userId = KeycloakUtil.getTokenUserId();
-		ShanoirEvent event = new ShanoirEvent(ShanoirEventType.IMPORT_DATASET_EVENT, importJob.getExaminationId().toString(), userId, "Starting import...", ShanoirEvent.IN_PROGRESS, 0f);
-		eventService.publishEvent(event);
-
-		if (importJob == null || importJob.getDatasets() == null || importJob.getDatasets().isEmpty()) {
-			event.setStatus(ShanoirEvent.ERROR);
-			event.setMessage("No datasets to create. Please check your EEG files");
-			event.setProgress(1f);
-			eventService.publishEvent(event);
-			return;
-		}
-
-		try {
-			DatasetAcquisition datasetAcquisition = new EegDatasetAcquisition();
-
-			// Get examination
-			Examination examination = examinationService.findById(importJob.getExaminationId());
-
-			datasetAcquisition.setExamination(examination);
-			datasetAcquisition.setAcquisitionEquipmentId(importJob.getAcquisitionEquipmentId());
-			datasetAcquisition.setRank(0);
-			datasetAcquisition.setSortingIndex(0);
-
-			List<Dataset> datasets = new ArrayList<>();
-			float progress = 0f;
-
-			for (EegDatasetDTO datasetDto : importJob.getDatasets()) {
-				progress += 1f / importJob.getDatasets().size();
-				event.setMessage("Dataset " + datasetDto.getName() + " for examination " + importJob.getExaminationId());
-				event.setProgress(progress);
-				eventService.publishEvent(event);
-				// Metadata
-				DatasetMetadata originMetadata = new DatasetMetadata();
-				originMetadata.setProcessedDatasetType(ProcessedDatasetType.NONRECONSTRUCTEDDATASET);
-				originMetadata.setDatasetModalityType(DatasetModalityType.EEG_DATASET);
-				originMetadata.setName(datasetDto.getName());
-				originMetadata.setCardinalityOfRelatedSubjects(CardinalityOfRelatedSubjects.SINGLE_SUBJECT_DATASET);
-
-				// Create the dataset with informations from job
-				EegDataset datasetToCreate = new EegDataset();
-
-				// DatasetExpression with list of files
-				DatasetExpression expression = new DatasetExpression();
-				expression.setCreationDate(LocalDateTime.now());
-				expression.setDatasetExpressionFormat(DatasetExpressionFormat.EEG);
-				expression.setDataset(datasetToCreate);
-
-				List<DatasetFile> files = new ArrayList<>();
-
-				// Set files
-				if (datasetDto.getFiles() != null) {
-
-					// Copy the data somewhere else
-					final String subLabel = SUBJECT_PREFIX + importJob.getSubjectName();
-					final String sesLabel = SESSION_PREFIX + importJob.getExaminationId();
-
-					final File outDir = new File(niftiStorageDir + File.separator + EEG_PREFIX + File.separator + subLabel + File.separator + sesLabel + File.separator);
-					outDir.mkdirs();
-
-					// Move file one by one to the new directory
-					for (String filePath : datasetDto.getFiles()) {
-
-						File srcFile = new File(filePath);
-						String originalNiftiName = srcFile.getAbsolutePath().substring(filePath.lastIndexOf('/') + 1);
-						File destFile = new File(outDir.getAbsolutePath() + File.separator + originalNiftiName);
-						Path finalLocation = null;
-						try {
-							finalLocation = Files.copy(srcFile.toPath(), destFile.toPath(), StandardCopyOption.REPLACE_EXISTING);
-						} catch (IOException e) {
-							LOG.error("IOException generating EEG Dataset Expression", e);
-						}
-
-						// Create datasetExpression => Files
-						if (finalLocation != null) {
-							DatasetFile file = new DatasetFile();
-							file.setDatasetExpression(expression);
-							file.setPath(finalLocation.toUri().toString());
-							file.setPacs(false);
-							files.add(file);
-						}
-					}
-				}
-
-				expression.setDatasetFiles(files);
-				datasetToCreate.setDatasetExpressions(Collections.singletonList(expression));
-
-				// set the dataset_id where needed
-				for (Channel chan : datasetDto.getChannels()) {
-					chan.setDataset(datasetToCreate);
-					chan.setReferenceType(ChannelType.EEG);
-					// Parse channel name to get its type
-					for (ChannelType type : ChannelType.values()) {
-						if (chan.getName().contains(type.name())) {
-							chan.setReferenceType(type);
-						}
-					}
-				}
-				for (Event eventToImport : datasetDto.getEvents()) {
-					eventToImport.setDataset(datasetToCreate);
-				}
-
-				// Fill dataset with informations
-				datasetToCreate.setChannelCount(datasetDto.getChannels() != null? datasetDto.getChannels().size() : 0);
-				datasetToCreate.setChannels(datasetDto.getChannels());
-				datasetToCreate.setEvents(datasetDto.getEvents());
-				datasetToCreate.setCreationDate(LocalDate.now());
-				datasetToCreate.setDatasetAcquisition(datasetAcquisition);
-				datasetToCreate.setOriginMetadata(originMetadata);
-				datasetToCreate.setUpdatedMetadata(originMetadata);
-				datasetToCreate.setSubjectId(importJob.getSubjectId());
-				datasetToCreate.setSamplingFrequency(datasetDto.getSamplingFrequency());
-				datasetToCreate.setCoordinatesSystem(datasetDto.getCoordinatesSystem());
-
-				datasets.add(datasetToCreate);
-			}
-
-			datasetAcquisition.setDatasets(datasets);
-			datasetAcquisitionService.create(datasetAcquisition);
-			
-			event.setProgress(1f);
-			event.setStatus(ShanoirEvent.SUCCESS);
-			// This message is important for email service
-			event.setMessage(importJob.getStudyName() + "(" + importJob.getStudyId() + ")"
-					+": Successfully created datasets for subject " + importJob.getSubjectName()
-					+ " in examination " + examination.getId());
-			eventService.publishEvent(event);
-
-			// Send mail
-			sendImportEmail(importJob, userId, examination, Collections.singleton(datasetAcquisition));
-		} catch (Exception e) {
-			LOG.error("Error while importing EEG: ", e);
-			event.setStatus(ShanoirEvent.ERROR);
-			event.setMessage("An unexpected error occured, please contact an administrator.");
-			event.setProgress(1f);
-			eventService.publishEvent(event);
-
-			// Send failure mail
-			sendFailureMail(importJob, userId, e.getMessage());
-			throw e;
-		}
-	}
 
 	/**
 	 * Create a processed dataset dataset associated with a dataset processing.
 	 * @param importJob the import job from importer MS.
 	 */
-	public void createProcessedDataset(final ProcessedDatasetImportJob importJob) {
+	public void createProcessedDataset(final ProcessedDatasetImportJob importJob) throws IOException {
 
-		ShanoirEvent event = new ShanoirEvent(ShanoirEventType.IMPORT_DATASET_EVENT, importJob.getProcessedDatasetFilePath().toString(), KeycloakUtil.getTokenUserId(), "Starting import...", ShanoirEvent.IN_PROGRESS, 0f);
+		ShanoirEvent event = new ShanoirEvent(ShanoirEventType.IMPORT_DATASET_EVENT, importJob.getProcessedDatasetFilePath(), KeycloakUtil.getTokenUserId(), "Starting import...", ShanoirEvent.IN_PROGRESS, 0f);
 		eventService.publishEvent(event);
 
 		if (importJob == null || importJob.getDatasetProcessing() == null) {
@@ -635,11 +354,12 @@ public class ImporterService {
 			DatasetFile datasetFile = new DatasetFile();
 			datasetFile.setPacs(false);
 			datasetFile.setPath(location.toUri().toString());
-			
+
 			DatasetExpression expression = new DatasetExpression();
 			expression.setDataset(dataset);
 			expression.setDatasetExpressionFormat(DatasetExpressionFormat.NIFTI_SINGLE_FILE);
 			expression.setDatasetProcessingType(datasetProcessing.getDatasetProcessingType());
+			expression.setSize(Files.size(location));
 			
 			datasetFile.setDatasetExpression(expression);
